+++ apps/game/src/prefs.h
@@ -5,7 +5,7 @@ ecs_comp_extern_public(GamePrefsComp) {
   bool dirty; // Indicates that the preference file should be saved to disk.
   f32  volume;
   bool fullscreen;
-  u32  windowWidth, windowHeight;
+  u16  windowWidth, windowHeight;
 };
 
 GamePrefsComp* prefs_init(EcsWorld*);
